Add verification checks + quality ratings to experiment pipeline
- experiment_splitter.py: verify_code_experiment() runs 3 checks
  (diff_sanity, syntax, import) and awards +0.5 quality bonus for
  passing. Shown inline during extraction.
- schema.py: quality field changed from int to float (allows 4.5)
- dream-forge.md: revised scale (5=verified, 4=real code, 3=decent)
- Raw sessions updated with Claude's quality ratings
- All 9 experiments regenerated with verified quality scores

+++ src/capture/schema.py
@@ -156,7 +156,7 @@ class Experiment(BaseModel):
     tags: list[str] = Field(default_factory=list)  # domain tags (python, android, web, etc.)
     confidence: ConfidenceTier = ConfidenceTier.INFERRED
     difficulty: Optional[str] = None  # "hard" | "easy" — Claude's estimate at capture time
-    quality: Optional[int] = None     # 1-5 — Claude's data quality rating (5=focused code fix with real context, 1=vague/broad)
+    quality: Optional[float] = None    # 1-5 — data quality rating (5=verified+focused, 4=real code unverified, 3=decent, 2=weak, 1=filler). Splitter adds +0.5 for passing verification checks.
 
     # Memory governance
     retrieval_count: int = 0
